Tests/LibWeb/Text/input/DOM/Document-methods-that-must-do-nothing.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        println(`document.clear() returns undefined: ${document.clear() === undefined}`);
        println(`document.captureEvents() returns undefined: ${document.captureEvents() === undefined}`);
        println(`document.releaseEvents() returns undefined: ${document.releaseEvents() === undefined}`);
    });
</script>
