Tests/LibWeb/Text/expected/DOM/Document-methods-that-must-do-nothing.txt

document.clear() returns undefined: true
document.captureEvents() returns undefined: true
document.releaseEvents() returns undefined: true
